Added IIOD client class
The IIOD client is supposed to be used by all backends which interface
the IIO Daemon. It will provide a private API that the backends can use
for most of the operations.

For now, only the function iiod_client_get_version() is provided. The
libiio backends can then wrap their own get_version() backend function
around this one.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>