IIOD: Fix bug introduced in commit 5b9c523
Before commit 5b9c523, reading an attribute from sysfs was likely to
return a string ending with \0, the \0 being counted as part of the
data (so 'abc\0' is a string of 4 characters).

A \n would then be appended in a second I/O call.

Commit 5b9c523 changed that, so that only one I/O call would be
performed. However, it replaced the data's \0 with a \n, which was
different from the original behaviour, and was breaking the ABI.

This commit returns to the original behaviour while still performing the
operations in one I/O call.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>