IIOD client: Fix communication error when reading mask
The iiod_client_read function was only using one I/O call to read the
mask, and then issuing an error if the number of bytes returned was
smaller than expected. This did not work well in practice; with the
network backend, the function would randomly return an error caused by a
short read of the mask.

Now, we make sure that the channel mask is read entirely.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>