Remove iio_create_usb_context_from_string()
Trying to guess which backend to use based on which delimiter characters
appear in the description string is not very scalable. In addition to that
iio_create_usb_context_from_string() does device identification based on
product and vendor ID, which does not work when multiple devices with the
same product and vendor ID are connected.

In light of this drop the recently introduced
iio_create_usb_context_from_string(), which has not been part of a release
yet, again. It will be replaced with a more scalable solution.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>